refactor: moved block to virtio_block
Now virtio_block is named similar to vhost_user_block
