+++ src/vmm/src/devices/virtio/mod.rs
@@ -11,7 +11,6 @@ use std::any::Any;
 use std::io::Error as IOError;
 
 pub mod balloon;
-pub mod block;
 pub mod block_common;
 pub mod device;
 pub mod gen;
@@ -24,6 +23,7 @@ pub mod rng;
 pub mod test_utils;
 pub mod vhost_user;
 pub mod vhost_user_block;
+pub mod virtio_block;
 pub mod vsock;
 
 /// When the driver initializes the device, it lets the device know about the
