+++ src/vmm/src/devices/virtio/persist.rs
@@ -241,13 +241,13 @@ mod tests {
     use utils::tempfile::TempFile;
 
     use super::*;
-    use crate::devices::virtio::block::device::FileEngineType;
-    use crate::devices::virtio::block::test_utils::default_block_with_path;
-    use crate::devices::virtio::block::VirtioBlock;
     use crate::devices::virtio::mmio::tests::DummyDevice;
     use crate::devices::virtio::net::test_utils::default_net;
     use crate::devices::virtio::net::Net;
     use crate::devices::virtio::test_utils::default_mem;
+    use crate::devices::virtio::virtio_block::device::FileEngineType;
+    use crate::devices::virtio::virtio_block::test_utils::default_block_with_path;
+    use crate::devices::virtio::virtio_block::VirtioBlock;
     use crate::devices::virtio::vsock::{Vsock, VsockUnixBackend};
 
     const DEFAULT_QUEUE_MAX_SIZE: u16 = 256;
