+++ src/vmm/src/lib.rs
@@ -132,8 +132,8 @@ use crate::devices::legacy::{IER_RDA_BIT, IER_RDA_OFFSET};
 use crate::devices::virtio::balloon::{
     Balloon, BalloonConfig, BalloonError, BalloonStats, BALLOON_DEV_ID,
 };
-use crate::devices::virtio::block::VirtioBlock;
 use crate::devices::virtio::net::Net;
+use crate::devices::virtio::virtio_block::VirtioBlock;
 use crate::devices::virtio::{TYPE_BALLOON, TYPE_BLOCK, TYPE_NET};
 use crate::logger::{error, info, warn, MetricsError, METRICS};
 use crate::persist::{MicrovmState, MicrovmStateError, VmInfo};
